VFB-311 rename filenames, rework override action case, add 'remove override' and 'remove all overrides' action reducer cases, separate out schema types, add further test coverage
